backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-multipart==0.0.9
numpy==1.26.4
rapidfuzz==3.9.7
faster-whisper==1.0.3
pypdfium2==4.30.0
python-docx==1.1.2
openpyxl==3.1.5
Pillow==10.4.0
pytesseract==0.3.13
anthropic==0.34.2
python-dotenv==1.0.1
aiofiles==24.1.0
